Refactor to ReflectionClassInstantiator with decorators.
